+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * next within a block can return a value
 * handle scans that return zero results with ESLevel < 2020
 * redo
+* add rand to filter functions
 
 # 4.0.1 / 2021-02-23